house controller

+++ house-biz/src/main/java/com/april/house/biz/mapper/CommentMapper.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface CommentMapper extends Mapper<Comment> {
 //    List<Comment> selectHouseComments(@Param("houseId") Long houseId, @Param("size") int size);
 
-    List<Comment> selectCommentsWithType(@Param("type") Integer type, @Param("typeId") Integer typeId, @Param("size") int size);
+    List<Comment> selectCommentsWithType(@Param("type") Integer type, @Param("typeId") Long typeId, @Param("size") int size);
 
 
 }
